feat:  setup returning dynamic types (#2)
We don't yet execute the queries using Postgres but we now have the capability to return different types (and names) based on what the input Substrait plan wants.

The AS clause is required for from_substrait and from_substrait_json calls but if an incorrect one is supplied the correct one is provided in the error message.